# perf(ivy): add benchmark for `[class]=exp` use case