This PR demonstrates a use case where writing to the `[class]` is much slower that direct.

```
Benchmark: class binding: create:
  [class]=" '1' ": 162.153 ns(0%)
  [class]=" '1 2' ": 209.617 ns(-29%)
  [class]=" '1 2 3 4 5 6 7 8 9 0' ": 213.104 ns(-31%)
  class="A B" [class]=" '1' ": 269.382 ns(-66%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' ": 242.590 ns(-50%)
  class="A B" [class]=" '1' " [class.foo]="exp": 241.773 ns(-49%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' " [class.foo]="exp": 242.958 ns(-50%)
  [element.class]="exp": 194.625 ns(-20%)
Benchmark: class binding: update:
  [class]=" '1' ": 17.339 ns(0%)
  [class]=" '1 2' ": 30.814 ns(-78%)
  [class]=" '1 2 3 4 5 6 7 8 9 0' ": 30.544 ns(-76%)
  class="A B" [class]=" '1' ": 30.543 ns(-76%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' ": 30.569 ns(-76%)
  class="A B" [class]=" '1' " [class.foo]="exp": 30.543 ns(-76%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' " [class.foo]="exp": 30.543 ns(-76%)
  [element.class]="exp": 30.543 ns(-76%)
Benchmark: class binding: noop:
  [class]=" '1' ": 16.239 ns(0%)
  [class]=" '1 2' ": 28.265 ns(-74%)
  [class]=" '1 2 3 4 5 6 7 8 9 0' ": 28.019 ns(-73%)
  class="A B" [class]=" '1' ": 28.111 ns(-73%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' ": 28.111 ns(-73%)
  class="A B" [class]=" '1' " [class.foo]="exp": 28.111 ns(-73%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' " [class.foo]="exp": 28.111 ns(-73%)
  [element.class]="exp": 28.110 ns(-73%)
```

This is first step in understanding the issue.