@pkozlowski-opensource 2019-10-23
@mhevery so the benchmark itself LGTM (in the sense of cases tested). What we should figure out, though, is presence of multiple cases in one benchmark and the `node packages/core/test/render3/perf/profile_all.js --write baseline.json` workflow.

Currently it seems to report only the first benchmark, here is the example output:

```
------------------------------------------------
PROFILING
------------------------------------------------
---------------- class_binding ----------------
Benchmark: class binding
  [class]=" '1' ": 326.413 us(-50%)
  [class]=" '1 2' ": 461.204 us(-111%)
  [class]=" '1 2 3 4 5 6 7 8 9 0' ": 1801.160 us(-726%)
  class="A B" [class]=" '1' ": 341.436 us(-57%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' ": 2112.785 us(-869%)
  class="A B" [class]=" '1' " [class.foo]="exp": 399.083 us(-83%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' " [class.foo]="exp": 2078.256 us(-853%)
  [element.class]="exp": 218.084 us(0%)
---------------- directive_instantiate ----------------
Benchmark: directive instantiate
  create: 2.569 us(0%)
---------------- element_text_create ----------------
Benchmark: element and text create
  create: 1.246 us(0%)
---------------- interpolation ----------------
Benchmark: interpolation refresh
  refresh: 227.722 us(0%)
---------------- listeners ----------------
Benchmark: listeners create
  create: 2.746 us(0%)
---------------- map_based_style_and_class_bindings ----------------
Benchmark: style and class map binding
  refresh: 20.180 ms(0%)
---------------- noop_change_detection ----------------
Benchmark: noop change detection
  refresh: 22.840 us(0%)
---------------- property_binding ----------------
Benchmark: property binding refresh
  refresh: 252.522 us(0%)
---------------- property_binding_update ----------------
Benchmark: property binding
  update: 456.747 us(0%)
---------------- style_and_class_bindings ----------------
Benchmark: style and class binding
  refresh: 1.089 ms(0%)
---------------- style_binding ----------------
Benchmark: style binding
  refresh: 481.473 us(0%)
================================================
┌────────────────────────────────────┬─────────┬──────┬───────────┬───────────┬───┐
│              (index)               │  time   │ unit │ base_time │ base_unit │ % │
├────────────────────────────────────┼─────────┼──────┼───────────┼───────────┼───┤
│           class_binding            │ 326.413 │ 'us' │           │           │   │
│       directive_instantiate        │  2.569  │ 'us' │           │           │   │
│        element_text_create         │  1.246  │ 'us' │           │           │   │
│           interpolation            │ 227.722 │ 'us' │           │           │   │
│             listeners              │  2.746  │ 'us' │           │           │   │
│ map_based_style_and_class_bindings │  20.18  │ 'ms' │           │           │   │
│       noop_change_detection        │  22.84  │ 'us' │           │           │   │
│          property_binding          │ 252.522 │ 'us' │           │           │   │
│      property_binding_update       │ 456.747 │ 'us' │           │           │   │
│      style_and_class_bindings      │  1.089  │ 'ms' │           │           │   │
│           style_binding            │ 481.473 │ 'us' │           │           │   │
└────────────────────────────────────┴─────────┴──────┴───────────┴───────────┴───┘
```

Notice that in the output table `class_binding` reports results of the first run (`[class]=" '1' "`). I presume that we either want to report all results or have the reported one being "the most representative". We can do this in the follow up PR, somehow I'm leaning towards reporting all cases, but would love to have your thoughts here.

.