@mhevery 2019-10-23
> @mhevery so the benchmark itself LGTM (in the sense of cases tested). What we should figure out, though, is presence of multiple cases in one benchmark and the `node packages/core/test/render3/perf/profile_all.js --write baseline.json` workflow.

I am aware of this. I will do a follow up PR to fix that issue. I wanted to re-do the whole thing anyways.