@mhevery 2019-10-24
Closing in favor of https://github.com/angular/angular/pull/33375